+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-08-11 15:03:53
-// Commit: chore(storage): update dependencies and cleanup unused imports in storage
+// Updated at: 2026-08-11 17:15:48
+// Commit: feat(storage): add async storage stream provider
 
 class storageComponentState {
-  final String commitRef = '3528';
-  final String lastMessage = 'chore(storage): update dependencies and cleanup unused imports in storage';
-  final DateTime timestamp = DateTime.parse('2026-08-11T15:03:53+05:00');
+  final String commitRef = '3537';
+  final String lastMessage = 'feat(storage): add async storage stream provider';
+  final DateTime timestamp = DateTime.parse('2026-08-11T17:15:48+05:00');
 }